# Retro-Forum-Hub/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: HiddenFreeways API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /chat/rooms:
    get:
      operationId: listChatRooms
      tags: [chat]
      summary: List chat rooms accessible to current user
      responses:
        "200":
          description: Rooms
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatRoom"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createChatRoom
      tags: [admin, chat]
      summary: Create a chat room (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateChatRoomBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatRoom"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /chat/rooms/{slug}/messages:
    get:
      operationId: listRoomMessages
      tags: [chat]
      summary: List recent messages for a chat room
      parameters:
        - in: path
          name: slug
          required: true
          schema: { type: string }
        - in: query
          name: sinceId
          required: false
          schema: { type: integer }
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoomMessagesResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Room not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: sendRoomMessage
      tags: [chat]
      summary: Send a chat message into a room
      parameters:
        - in: path
          name: slug
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendChatMessageBody"
      responses:
        "201":
          description: Sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoomMessage"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ------------------- CREWS -------------------
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]

    # ---------- AUTH ----------
...
    ChatRoom:
      type: object
      properties:
        id:
          type: integer
        slug:
          type: string
        name:
          type: string
        description:
          type: string
        kind:
          type: string
          enum: [public, state, trusted]
        stateId:
          type: ["integer", "null"]
        stateSlug:
          type: ["string", "null"]
        minTrustLevel:
          type: integer
        isArchived:
          type: boolean
        memberCount:
          type: integer
        lastMessageAt:
          type: ["string", "null"]
          format: date-time
      required:
        - id
        - slug
        - name
        - description
        - kind
        - stateId
        - stateSlug
        - minTrustLevel
        - isArchived
        - memberCount
        - lastMessageAt
    CreateChatRoomBody:
      type: object
      properties:
        slug:
          type: string
        name:
          type: string
        description:
          type: string
        kind:
          type: string
          enum: [public, state, trusted]
        stateId:
          type: ["integer", "null"]
        minTrustLevel:
          type: integer
      required: [slug, name, description, kind]
    RoomMessage:
      type: object
      properties:
        id:
          type: integer
        body:
          type: string
        authorId:
          type: integer
        authorUsername:
          type: string
        authorTrustLevel:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - body
        - authorId
        - authorUsername
        - authorTrustLevel
        - createdAt
    RoomMessagesResponse:
      type: object
      properties:
        messages:
          type: array
          items:
            $ref: "#/components/schemas/RoomMessage"
      required: [messages]
    SendChatMessageBody:
      type: object
      properties:
        body:
          type: string
      required: [body]
...